+++ showtime-squad_frontend/src/components/atoms/FunctionButton.scss
@@ -1,7 +1,7 @@
 .function-button-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+    // display: flex;
+    // flex-direction: column;
+    // align-items: center;
     
     .error {
         position: absolute;
