+++ showtime-squad_frontend/src/pages/Groups/GroupView.jsx
@@ -1,47 +1,6 @@
-import { Routes, Route } from "react-router-dom"
-import { signal } from '@preact/signals-react'
 
-import { deleteRequest, getRequest, postRequest } from "../../utils/GenericHTTPMethods"
-import { useUser } from '../../context/UserContext';
 
-const apiUrl = import.meta.env.VITE_REACT_APP_BACKEND_BASE_URL
-
-function GroupView({ name = "", showSignal, groupSignal }) {
-    if (showSignal == null || groupSignal == null) {
-        throw new Error("Specified (individual) signals are mandatory")
-    }
-
-    const { username } = useUser()
-
-    async function handleShow() {
-        showSignal.value = !showSignal.value
-        if (showSignal.value === false) {
-            return
-        }
-
-        const group = await fetchGroupInfo(name, username)
-        if (!group) {
-            console.error("Could not fetch group")
-            return;
-        }
-
-        console.log(group)
-        groupSignal.value = group
-    }
-
-    return (
-        <>
-            <section>
-                {name} - {showSignal.value ? 'show please' : 'dont show'}
-                <button onClick={handleShow}>Click me</button>
-
-                {showSignal.value ? <Idklol group={groupSignal.value} /> : <></>}
-            </section>
-        </>
-    )
-}
-
-function Idklol({ group }) {
+function GroupView({ group }) {
     if (!group) {
         return (
             <>
@@ -82,44 +41,6 @@ function Idklol({ group }) {
             ) : <></>}
         </>
     )
-}
-
-async function fetchGroupInfo(name, username) {
-    try {
-        const response = await getRequest({ url: `${apiUrl}/api/group/${name}` })
-        if (response && response.groupname) {
-            // response seems valid
-            if (username && response.owner && username == response.owner) {
-                //is owner, get join requests
-                console.log("is owner!")
-                const joiners = await postRequest({
-                    url: `${apiUrl}/api/group/requests`,
-                    body: { groupname: response.groupname }
-                });
-
-                console.log("joiners: ", joiners)
-                if (joiners) {
-                    response.joinRequests = joiners.joinRequests;
-                } else {
-                    throw new Error("Could not fetch joiners")
-                }
-            }
-
-            if(response.news && response.news.length > 0) {
-                // fetch news
-            }
-
-            console.log("____", response)
-
-            return response;
-        }
-
-        return null;
-    } catch (error) {
-        console.error("Error fetching groups:", error);
-        // Handle error if fetchGroups fails
-        return null;
-    }
 }
 
 export default GroupView
